Separate handling of error output from generation in Linter CLI
